# Optimize truth value testing for strings