This causes code that tests for the boolean value of a string to test the length directly instead of going through PyObject_IsTrue, as mentioned in mypyc/mypyc#768.

This should make code that uses this faster, but my benchmarks seem to show that this makes code slower than it would be without this optimization. I was hoping you could take a look at my code and the benchmarks and see if I'm doing anything wrong.

## Test Plan

I added a test that compiles and runs code that uses this, as well as an irbuild test, and ran all of the mypyc tests to make sure they were still working.

## Benchmarks

This is my benchmark code:
```python
import time

def run(s: str) -> None:
    t0 = time.time()
    for i in range(1_000_000):
        if s:
            pass
    t1 = time.time()
    print(f'{t1 - t0:.3f}')

run('')
run('A' * 100)
```

My results from running this benchmark several times each:

Interpreted version without mypyc: 0.022 to 0.024 seconds
Compiled version, from master (without these changes): 0.006 to 0.007 seconds
Compiled version, with these changes: 0.008 to 0.009 seconds