@pranavrajpal 2021-03-31
Ok, I have moved the check for whether a string's length is 0 to a primitive, and this appears to have made it faster.

Using the previously mentioned benchmark, I got these times:
Interpreted: 0.022 to 0.024 seconds
Compiled version, without these changes: 0.006 to 0.007 seconds
Compiled version, with these changes: 0.003 to 0.004 seconds

That appears to show that these changes make that benchmark approximately twice as fast.